test: move tests to tests folder
